rename otpview
