+++ src/v2/view-builder/views/email/RequiredFactorEmailView.js
